+++ packages/ui/VirtualList/VirtualListBase.js
@@ -319,8 +319,9 @@ const VirtualListBaseFactory = (type) => {
 
 		componentDidUpdate (prevProps, prevState) {
 			let deferScrollTo = false;
-
 			const {firstIndex, numOfItems} = this.state;
+
+			this.shouldUpdateBounds = false;
 
 			// TODO: remove `this.hasDataSizeChanged` and fix ui/Scrollable*
 			this.hasDataSizeChanged = (prevProps.dataSize !== this.props.dataSize);
@@ -432,6 +433,8 @@ const VirtualListBaseFactory = (type) => {
 
 		isPrimaryDirectionVertical = true
 		isItemSized = false
+
+		shouldUpdateBounds = false
 
 		dimensionToExtent = 0
 		threshold = 0
@@ -615,6 +618,12 @@ const VirtualListBaseFactory = (type) => {
 				wasFirstIndexMax = ((this.maxFirstIndex < moreInfo.firstVisibleIndex - dimensionToExtent) && (firstIndex === this.maxFirstIndex)),
 				dataSizeDiff = dataSize - this.curDataSize;
 			let newFirstIndex = firstIndex;
+
+			// When calling setState() except in didScroll(), `shouldUpdateBounds` should be `true`
+			// so that setState() in didScroll() could be called to override state.
+			// Before calling setState() except in didScroll(), getStatesAndUpdateBounds() is always called.
+			// So `shouldUpdateBounds` is true here.
+			this.shouldUpdateBounds = true;
 
 			this.maxFirstIndex = Math.ceil((dataSize - numOfItems) / dimensionToExtent) * dimensionToExtent;
 			this.curDataSize = dataSize;
@@ -883,7 +892,9 @@ const VirtualListBaseFactory = (type) => {
 			this.scrollPosition = pos;
 			this.updateMoreInfo(dataSize, pos);
 
-			this.setState({firstIndex: newFirstIndex});
+			if (this.shouldUpdateBounds || firstIndex !== newFirstIndex) {
+				this.setState({firstIndex: newFirstIndex});
+			}
 		}
 
 		// For individually sized item